+++ src/main/java/me/yleoft/zHomes/commands/MainVersionSubCommand.java
@@ -10,10 +10,6 @@ import java.util.List;
 
 public class MainVersionSubCommand implements SubCommand {
 
-    public MainVersionSubCommand() {
-        addSubCommand(new MainVersionUpdateSubCommand());
-    }
-
     @Override
     public @NotNull String name() {
         return "version";
@@ -39,50 +35,10 @@ public class MainVersionSubCommand implements SubCommand {
                 message(sender, "%prefix% <gold>You can update your plugin here: <yellow>" + UpdateUtils.site);
             }else {
                 message(sender, zHomes.getLanguageYAML().getMainVersionOutput());
-                message(sender, zHomes.getLanguageYAML().getMainVersionNoUpdate());
             }
             return;
         }
         message(sender, zHomes.getLanguageYAML().getMainVersionOutput());
-    }
-
-    @Override
-    public @NotNull List<String> tabComplete(@NotNull CommandSender sender, @NotNull String[] fullArgs, @NotNull String[] args) {
-        return sender.hasPermission(zHomes.getConfigYAML().getMainCommandVersionUpdatePermission()) ?
-                List.of("update") : List.of();
-    }
-
-    public static class MainVersionUpdateSubCommand implements SubCommand {
-
-        @Override
-        public @NotNull String name() {
-            return "update";
-        }
-
-        @Override
-        public List<String> aliases() {
-            return List.of("--update");
-        }
-
-        @Override
-        public String permission() {
-            return zHomes.getConfigYAML().getMainCommandVersionUpdatePermission();
-        }
-
-        @Override
-        public void execute(@NotNull CommandSender sender, @NotNull String[] fullArgs, @NotNull String @NotNull [] args) {
-            if(!zHomes.updateUtils.needsUpdate) {
-                message(sender, zHomes.getLanguageYAML().getMainVersionNoUpdate());
-                return;
-            }
-            try {
-                zHomes.updateUtils.checker.update();
-                message(sender, zHomes.getLanguageYAML().getMainVersionUpdateOutput());
-            } catch (Exception e) {
-                zHomes.getInstance().getLoggerInstance().warn("An error occurred while updating the plugin", e);
-            }
-        }
-
     }
 
 }